Move QueryPlanningContext into its own file
`QueryPlanningContext` is its own class but it was inline in
`buildQueryPlan` so other modules that are used as part of
`buildQueryPlan` couldn't depend on the context without creating
circular dependencies (plus, this made `buildQueryPlan.ts` longer than
necesary).

This commit moves the `QueryPlanningContext` class in its own
module. For that, it also moves the `collectFields` method as a
method of `QueryPlanningContext`, since it is used within some of
the context methods.

There should be no functional changes otherwise.
